requirements.txt
fastapi==0.96.0
uvicorn==0.22.0
sqlalchemy==2.0.17
pydantic[email]==1.10.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
alembic==1.11.1
psycopg2-binary==2.9.6
